+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - 修复: windows x64平台调用npminstall失败
 - 修复: 窗口崩溃时间存储错误
 - 修复: T3Drpy和T3Catopen适配器同步方法postType为form请求错误; T0Xml和T1Json适配器处理playUrl获取链接错误
+- 修复: 同步请求内容非标准json格式化错误(json5替代); axios子进程请求支持代理; axios请求头返回不标准
 - 修复: 老板键触发缺未暂停视频播放
 - 修复: cms接口返回url数组时无法正确处理
 - 修复: 影视有源但为选择时无法触发全部搜索问题